+++ src/agent/samplers/syscall/linux/counts/mod.rs
@@ -18,7 +18,6 @@ use stats::*;
 
 use super::syscall_lut;
 use crate::agent::*;
-use crate::*;
 
 use std::sync::Arc;
