+++ src/agent/samplers/tcp/linux/connect_latency/mod.rs
@@ -19,7 +19,6 @@ use bpf::*;
 use stats::*;
 
 use crate::agent::*;
-use crate::*;
 
 use std::sync::Arc;
