+++ src/agent/samplers/tcp/linux/packet_latency/mod.rs
@@ -18,7 +18,6 @@ use bpf::*;
 use stats::*;
 
 use crate::agent::*;
-use crate::*;
 
 use std::sync::Arc;
