+++ src/agent/samplers/tcp/linux/receive/mod.rs
@@ -17,7 +17,6 @@ use bpf::*;
 use stats::*;
 
 use crate::agent::*;
-use crate::*;
 
 use std::sync::Arc;
