+++ src/agent/samplers/tcp/linux/retransmit/mod.rs
@@ -16,7 +16,6 @@ use bpf::*;
 use stats::*;
 
 use crate::agent::*;
-use crate::*;
 
 use std::sync::Arc;
