+++ src/agent/samplers/tcp/linux/traffic/mod.rs
@@ -22,7 +22,6 @@ use bpf::*;
 use stats::*;
 
 use crate::agent::*;
-use crate::*;
 
 use std::sync::Arc;
